requirements.txt
paho-mqtt==1.6.1
keybert==0.7.0
transformers==4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
tf-keras